requirements.txt
fastapi
uvicorn
streamlit
influxdb-client
google.generativeai        # Replace with other LLM libraries if necessary
